+++ Tests/HandySwiftTests/Extensions/CollectionExtensionTests.swift
@@ -28,7 +28,13 @@ class CollectionExtensionTests: XCTestCase {
         let intArray = [1, 2, 10]
         XCTAssertEqual(intArray.average(), 4.333, accuracy: 0.001)
 
+        let averageAsCGFloat: CGFloat = intArray.average()
+        XCTAssertEqual(averageAsCGFloat, 4.333, accuracy: 0.001)
+
         let doubleArray = [1.0, 2.0, 10.0]
         XCTAssertEqual(doubleArray.average(), 4.333, accuracy: 0.001)
+
+        let cgFloatArray: [CGFloat] = [1.0, 2.0, 10.0]
+        XCTAssertEqual(cgFloatArray.average(), 4.333, accuracy: 0.001)
     }
 }